Vamped out delete command
Took 6 minutes
